requirements.txt
pandas
jupyter
pyhwp
